Added healthcheck for mysql container
Healthcheck is supported since docker-compose version 3, this was
released back in 2017 so it's a safe bet that everyone doing this course
will be on a supported version.
To check if mysql is accepting connections a shell will be used with the
username and password given in the environment, this will run the
command SHOW DATABASES. mysql will return exit code 1 if there is a
problem while executing and 0 if it succeeded.